ci: also run linux-gcc pipeline with python3.5 environment
git-p4.py includes support for python3, but this was not previously
validated in CI. Lets actually do that.

As of writing, python-3.5 has reached end-of-life, but has been updated
recently enough that it's reasonable to attemp to support it. We do not
have a pressing need for features only available in 3.6 and later.

Usage of python3 is limited to the linux-gcc pipeline on Azure. It is
assumed that passing both python2 and python3 tests on one platform
translates to doing the same on others.

Travis-CI is unchanged, as no tests are run in those environments.

Signed-off-by: Yang Zhao <yang.zhao@skyboxlabs.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>